move code chunks from commands.rs to lib.rs for reuse
